+++ changelog.md
@@ -1,1 +1,2 @@
-- Updated to 1.18
+- Fixed grave not saving data
+- Re added Jade support